reduce to just next / map
