+++ tokio/src/signal/windows.rs
@@ -245,7 +245,8 @@ pub fn ctrl_break() -> io::Result<CtrlBreak> {
 #[cfg(all(test, not(loom)))]
 mod tests {
     use super::*;
-    use crate::{runtime::Runtime, stream::StreamExt};
+    use crate::runtime::Runtime;
+    use crate::stream::StreamExt;
 
     use tokio_test::{assert_ok, assert_pending, assert_ready_ok, task};
