--- tokio/src/stream/streamext/filter.rs
@@ -1,127 +1,0 @@
-use crate::stream::Stream;
-
-use core::fmt;
-use core::future::Future;
-use core::pin::Pin;
-use core::task::{Context, Poll};
-use pin_project_lite::pin_project;
-
-pin_project! {
-    /// Stream for the [`filter`](super::StreamExt::filter) method.
-    #[must_use = "streams do nothing unless polled"]
-    pub struct Filter<St, Fut, F>
-        where St: Stream,
-    {
-        #[pin]
-        stream: St,
-        f: F,
-        #[pin]
-        pending_fut: Option<Fut>,
-        pending_item: Option<St::Item>,
-    }
-}
-
-impl<St, Fut, F> fmt::Debug for Filter<St, Fut, F>
-where
-    St: Stream + fmt::Debug,
-    St::Item: fmt::Debug,
-    Fut: fmt::Debug,
-{
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("Filter")
-            .field("stream", &self.stream)
-            .field("pending_fut", &self.pending_fut)
-            .field("pending_item", &self.pending_item)
-            .finish()
-    }
-}
-
-impl<St, Fut, F> Filter<St, Fut, F>
-where St: Stream,
-      F: FnMut(&St::Item) -> Fut,
-      Fut: Future<Output = bool>,
-{
-    pub(super) fn new(stream: St, f: F) -> Filter<St, Fut, F> {
-        Filter {
-            stream,
-            f,
-            pending_fut: None,
-            pending_item: None,
-        }
-    }
-
-    /// Acquires a reference to the underlying stream that this combinator is
-    /// pulling from.
-    pub fn get_ref(&self) -> &St {
-        &self.stream
-    }
-
-    /// Acquires a mutable reference to the underlying stream that this
-    /// combinator is pulling from.
-    ///
-    /// Note that care must be taken to avoid tampering with the state of the
-    /// stream which may otherwise confuse this combinator.
-    pub fn get_mut(&mut self) -> &mut St {
-        &mut self.stream
-    }
-
-    /// Acquires a pinned mutable reference to the underlying stream that this
-    /// combinator is pulling from.
-    ///
-    /// Note that care must be taken to avoid tampering with the state of the
-    /// stream which may otherwise confuse this combinator.
-    pub fn get_pin_mut(self: Pin<&mut Self>) -> Pin<&mut St> {
-        self.project().stream
-    }
-
-    /// Consumes this combinator, returning the underlying stream.
-    ///
-    /// Note that this may discard intermediate state of this combinator, so
-    /// care should be taken to avoid losing resources when this is called.
-    pub fn into_inner(self) -> St {
-        self.stream
-    }
-}
-
-impl<St, Fut, F> Stream for Filter<St, Fut, F>
-    where St: Stream,
-          F: FnMut(&St::Item) -> Fut,
-          Fut: Future<Output = bool>,
-{
-    type Item = St::Item;
-
-    fn poll_next(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-    ) -> Poll<Option<St::Item>> {
-        loop {
-            if self.pending_fut.is_none() {
-                let item = match ready!(self.as_mut().project().stream.poll_next(cx)) {
-                    Some(e) => e,
-                    None => return Poll::Ready(None),
-                };
-                let fut = (self.as_mut().project().f)(&item);
-                self.as_mut().project().pending_fut.set(Some(fut));
-                *self.as_mut().project().pending_item = Some(item);
-            }
-
-            let yield_item = ready!(self.as_mut().project().pending_fut.as_pin_mut().unwrap().poll(cx));
-            self.as_mut().project().pending_fut.set(None);
-            let item = self.as_mut().project().pending_item.take().unwrap();
-
-            if yield_item {
-                return Poll::Ready(Some(item));
-            }
-        }
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        let pending_len = if self.pending_item.is_some() { 1 } else { 0 };
-        let (_, upper) = self.stream.size_hint();
-        let upper = match upper {
-            Some(x) => x.checked_add(pending_len),
-            None => None,
-        };
-        (0, upper) // can't know a lower bound, due to the predicate
-    }
-}
